feat: add ownership validation in delete game use case

+++ game-list-api/src/domain/use-cases/games/delete-games.ts
@@ -1,14 +1,17 @@
 import { GameNotFoundError } from '@/domain/errors/game-not-found';
 import { PrismaGameRepository } from '@/infra/database/prisma/repositories/prisma-games-repository';
+import { ensureOwnership } from '@/utils/ensure-ownership';
 
 export class DeleteGameUseCase {
   constructor(private gameRepository: PrismaGameRepository) {}
 
-  async execute(id: string) {
+  async execute(id: string, userId: string) {
     const game = await this.gameRepository.findById(id);
     if (!game) {
       throw new GameNotFoundError();
     }
+
+    ensureOwnership(game.userId ?? '', userId);
 
     await this.gameRepository.delete(id);
   }
